+++ kugl/util/__init__.py
@@ -3,7 +3,7 @@ from .age import Age, parse_age, to_age
 from .clock import UNIT_TEST_TIMEBASE
 from .debug import debug_features, debugging, features_debugged
 from .misc import fail, KuglError, parse_utc, run, TABLE_NAME_RE, to_utc, warn, WHITESPACE_RE, cleave
-from .paths import KPath, ConfigPath, kugl_home, kube_home, kube_context
+from .paths import KPath, ConfigPath, kugl_home, kube_home, kugl_cache, kube_context
 from .size import parse_size, to_size, parse_cpu
 from .sqlite import SqliteDb
 from .sqlparse import Query