+++ tests/test_misc.py
@@ -3,8 +3,8 @@ Assorted tests for various edge cases and error conditions.
 Some of these are just to achieve 100% coverage.
 """
 
-import os
 from pathlib import Path
+import shutil
 from unittest.mock import patch
 
 import pytest
@@ -29,16 +29,16 @@ def test_limits_misc(capsys):
 
 
 @patch.dict(kube_context.memory, clear=True)  # suppress memoization
-def test_kube_home_missing(test_home, tmp_path):
-    os.environ["KUGL_HOME"] = str(tmp_path / "doesnt_exist")
+def test_kube_home_missing(test_home):
+    shutil.rmtree(str(kube_home()))
     with pytest.raises(KuglError, match="can't determine current context"):
-        # Must actually query a resource or KubernetesResource won't ask for the context
+        # Must actually query resource or KubernetesResource won't ask for the context
         main1(["select * from nodes"])
 
 
 @patch.dict(kube_context.memory, clear=True)  # suppress memoization
 def test_no_kube_context(test_home, tmp_path):
-    kube_home().joinpath("config").write_text("")
+    kube_home().prep().joinpath("config").write_text("")
     with pytest.raises(KuglError, match="No current context"):
         # Must actually query a resource or KubernetesResource won't ask for the context
         main1(["select * from nodes"])
@@ -51,8 +51,8 @@ def test_enforce_mockdir(test_home, monkeypatch):
 
 
 def test_kube_home_without_envar(monkeypatch):
-    monkeypatch.setenv("KUGL_HOME", "xxx")  # must exist before deleting
-    monkeypatch.delenv("KUGL_HOME")
+    monkeypatch.setenv("KUGL_KUBE_HOME", "xxx")  # must exist before deleting
+    monkeypatch.delenv("KUGL_KUBE_HOME")
     assert kube_home() == Path.home() / ".kube"
 
 
@@ -63,7 +63,7 @@ def test_kugl_home_without_envar(monkeypatch):
 
 
 def test_reject_world_writeable_config(test_home):
-    init_file = kugl_home() / "init.yaml"
+    init_file = kugl_home().prep() / "init.yaml"
     init_file.write_text("foo: bar")
     init_file.chmod(0o777)
     with pytest.raises(KuglError, match="is world writeable"):