+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 - Support multiple schemas & join across them
 - Allow comments for user-defined columns
 - Allow environment variables in `file` resource paths
-- Resource cache paths and file formats have changed, old `~/.kugl/cache` files are not compatible
-- `rm -r ~/.kugl/cache` is recommended before installing
+- Resource cache paths and file formats have changed, and cache now lives in `~/.kuglcache`
+- `rm -r ~/.kugl/cache` is recommended to clear obsolete files
 
 ## 0.3.3